Integrate ParallaxDataView for usage in Korge-fleks
Forked ImageAnimationView since some chagnes are neede for Korge-fleks
which are not yet upstreamed to Korge main repository.

In ParallaxDataView all update methods where removed since the view will
be updated by KorgeViewSystem of the Fleks ECS. That makes the current
state of the parallax view serializable and restorable from a save game.